Passage au long
devrait être suffisant pour les fitness, une branche utilisant BigInt
est conservée

+++ src/eu/labrush/agenetic/numfun/Fellow.java
@@ -2,8 +2,6 @@ package eu.labrush.agenetic.numfun;
 
 import eu.labrush.agenetic.AbstractFellow;
 
-import java.math.BigInteger;
-
 public class Fellow extends AbstractFellow {
 
     static private NumFunction fun ;
@@ -17,8 +15,8 @@ public class Fellow extends AbstractFellow {
         super(dna, DNACARD);
     }
 
-    public BigInteger calcFitness() {
-        return BigInteger.valueOf(fun.getY(this.cloneDNA()));
+    public long calcFitness() {
+        return fun.getY(this.cloneDNA());
     }
 
     public String getX(){
